+++ ChestsAnywhere/Menus/Overlays/ManageChestOverlay.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Linq;
-using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
@@ -10,7 +9,6 @@ using Pathoschild.Stardew.Common;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Menus;
-using Chest = StardewValley.Objects.Chest;
 
 namespace Pathoschild.Stardew.ChestsAnywhere.Menus.Overlays
 {
@@ -173,62 +171,7 @@ namespace Pathoschild.Stardew.ChestsAnywhere.Menus.Overlays
         /// <summary> Stacks items in the player's inventory with others found in the chests.</summary>
         public void StackItemsInInventory()
         {
-            // Combine an item stack with another of the same type.
-            //
-            // Returns whether the item stack has been completely combined.
-            Func<Item, Func<Item, bool>> combineItems = item => itemInChest => {
-                int space = item.getStack();
-                int availableSpace = Math.Min(space, itemInChest.getRemainingStackSpace());
-
-                if (availableSpace > 0)
-                {
-                    itemInChest.addToStack(availableSpace);
-                    item.addToStack(-1 * availableSpace);
-                }
-
-                return Math.Max(0, space - availableSpace) == 0;
-            };
-
-            // Index all available items in all chests for O(1) access later.
-            var itemsInChests = this.Chests
-                .Aggregate(new List<Item>{}, (list, managedChest) => {
-                    return list.Concat(managedChest.Chest.items).ToList();
-                })
-                .Aggregate(new Dictionary<string, List<Item>>{}, (map, item) => {
-                    if (!map.ContainsKey(item.Name))
-                    {
-                        map.Add(item.Name, new List<Item>{});
-                    }
-
-                    map[item.Name].Add(item);
-
-                    return map;
-                })
-            ;
-
-            // Player inventory items; the check is necessary because empty items are included
-            var items = Game1.player.items.Where(x => x != null);
-
-            // Items that have an existing stack in some of the chests
-            var eligibleItems = items.Where(x => itemsInChests.ContainsKey(x.Name));
-
-            var moves = eligibleItems.Select(item => {
-                return new {
-                    item,
-                    // keep combining the stack until it is drained, if possible:
-                    drained = itemsInChests[item.Name].Any(combineItems(item))
-                };
-            });
-
-            // Remove the items that were completely stacked from the player's inventory.
-            moves
-                .Where(move => move.drained)
-                .Select(move => move.item)
-                .ToList()
-                .ForEach(item => Game1.player.removeItemFromInventory(item))
-            ;
-
-            Game1.playSound("Ship");
+            ChestInventoryStacker.StackItemsInInventory(this.Chests.ToList());
         }
 
         /// <summary>Switch to the specified chest.</summary>
@@ -412,8 +355,6 @@ namespace Pathoschild.Stardew.ChestsAnywhere.Menus.Overlays
                         this.OpenEdit();
                     else if (input.Equals(config.SortItems))
                         this.SortInventory();
-                    else if (input.Equals(config.StackItems))
-                        this.StackItemsInInventory();
                     else
                         return false;
                     return true;
